Enable scraping of data on linux.
This Notebook serves the purpose to provide the tools to do scraping on my linux machine.
It enables me to collect data with my main windows laptop, but also with my small but powerfull linux dekstop.

It is also the first step to running automatized tasks on my linux machine using crontab to automatically scrape the data weekly.